password_cracking_by_using_dictionary_attack
How Password Are Hacked ?? EXPLANATION GIVEN BELLOW ?? 
..
..
but I will suggest try to solve it by own if not then go for it..
.
.
.
.
The explaination is :: --

Everyone knows that passwords are not stored a plain text in the website’s database. Now we are going to see how to hack a plain text password when you find a password that is in hashed(md5) format. So we take the input_hash(hashed password in the database) and try to compare it with md5 hash of every plain text password which is in a password file(pass_doc) and when the hashes are matched we simply display the plain text password which is in the password file(pass_doc). If the password is not present in the input password file it will say password is not found, this happens only if buffer overflow doesn’t occur. This type of attack can be considered as a dictionary attack.